Todo-list

+++ my-app/src/components/TodoItem.tsx
@@ -17,7 +17,9 @@ const TodoItem: FC<TodoItemProps> = (props) => {
     }
 
     const handleOnClick = () => {
-        dispatch(removeTodo(id))
+        if (window.confirm('remove todo?')){
+            dispatch(removeTodo(id))
+        }
     }
     return (
         <div>
